Librarian Books Search
Added functionality that allows the librarian to search for a book by book name or author.
Changed the book delete function to now mark books as inactive instead of removing them from the database.
Changed how an author age is saved on the database.  An authors date of birth will now be saved instead of their age.  Their age will now be calculated during table creation.