Added New Table to Database
Added a status table for rented books.
Changed the necessary functions to be inline with the newly added table.
Updated the headers to now display the logged in user's name and surname.
Added additional screens:
*Book's rental history
*Member's rental history
These screen can be viewed but the librarian.
Added a function that will update an active rental order's status to 'Overdue' if the return date has passed.
Added functionality to calculate a return date for when a member rents a book.