Password Hashing and Search Screens Update
Added password hashing when creating or updating an user's account.
Added hashed password verification when an user attempts to sign in.
Updated all search screens to be inline with the screen on which the search was performed.
Added a home button for the librarian to easily return to their main screen.